microservices/requirements.txt
Flask==2.0.1
mysql-connector-python==8.0.26
bcrypt==3.2.0
Flask-CORS==3.0.10
redis
pika==1.0.0
requests
apscheduler==3.10.4
